README: Clarifying what this fork is
This explains the reasoning for the creation of this fork, as well as adding a few links to relevant discussions.
There are probably more, but those are the few that I thought about.

This also makes a distinction between the "original" version, form the Audacity team, and the "modified" version provided in this repo. I thought it didn't really make sense to link to the original download links without at least clarifying this isn't the version being developed here...

Signed-off-by: Emily Mabrey <emilymabrey93@gmail.com>