Removed code signing and notarization from Windows/macOS builds
The scripts in the `scripts` directory are flexible, in the sense that they
were also developed with scenarios, such as not being supplied any credentials
whatsoever. This means that we can just remove the environment variables
concerning code, and it should just work.

These changes have not been tested thoroughly yet, but are an improvement
given our situation. We hope to bring these back eventually, but they cost
money.