TranslatableString: Make compatible with wx3.0
wxWidgets 3.0 doesn't support context in wxGetTranslation yet, so don't
use the context, similar to existing DoChooseFormat handling.
This will lead to some untranslated strings or wrong context translations
(if we have equal strings that one specifies the context, but the other
doesn't), which isn't great, but better than not being able to compile
the application at all against a stable wxWidgets library version.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>