Fixed Audacity.Project regression
What I tried to do here, as well as the alternative fixes I considered,
is way too large to explain. Long story short, I wanted to initially check
whether Tenacity.Project OR Audacity.Project was defined, but a typo as I was
formatting the change led to a logic error.

However, even if I did what I wanted to do, that would mean that we'd "lose"
our opportunity to ask the user whether they want to use Tenacity instead of
other programs. The Audacity developers basically already fixed the problem
that I was trying to fix, and the solution was right in front of us.

Signed-off-by: Panagiotis Vasilopoulos <hello@alwayslivid.com>