BUILDING.md: Add PortSMF Debian build instructions
The build instructions for PortSMF are missing from Debian section as it is
not included in the dependency list since the package is outdated and might
cause build problems during compilation. This commit adds instructions that
recommend the forked version of PortSMF to be used when possible (as it is
confirmed to work when building Tenacity).

Some additional changes to BUILDING.md were also made, such as some rewording
and the addition of a new "Notes" section for Debian/Ubuntu in order to organize
the information that the developer working on these platforms better.

Co-authored-by: Be <be.0@gmx.com>
Co-authored-by: Panagiotis Vasilopoulos <hello@alwayslivid.com>
Signed-off-by: Be <be.0@gmx.com>
Signed-off-by: Juozas <juozaspo@gmail.com>
Signed-off-by: Panagiotis Vasilopoulos <hello@alwayslivid.com>