Decide default theme on Windows based on the system theme
wxWidgets provides an easy way to read registry values, apparently.
Instead of forcing the user to switch their theme so that it corresponds
to the system theme, it'd be a better idea to just do that for them.

What this change does not do: If the user changes the system theme after
Tenacity is initially configured, the theme won't change. Some additional
style-related changes were also made.

Signed-off-by: Panagiotis Vasilopoulos <hello@alwayslivid.com>