Git is now a mandatory build dependency
We are planning to use Git to detect the current version of Tenacity, instead
of depending on a hardcoded variable, as that is an approach that is far from
foolproof. We already depended on it for Windows and macOS builds, as vcpkg
uses Git to fetch a couple of dependencies. The vcpkg has also been
appropriately adjusted and cleaned up.

It may seem a bit excessive, but the operating systems we support all have
a version of Git that is available and it is not a runtime dependency. We
assume that people who can gain access to the source code can also install
a copy of Git. If we actually come across some sort of an extreme edge case
where this is not true, we can set the versions to `9999` just like in
`win/tenacity.rc`, in order to ensure portability, at a later point in time.

This change is a measure that will ensure that we are going to avoid mistakes
that we will not be able to undo later.

Signed-off-by: Panagiotis Vasilopoulos <hello@alwayslivid.com>