Fix version detection in CMake build
The CMake build uses git describe to retrieve version information from
the latest git tag. However, if there is a tag like "alpha" which does
not match the expected pattern "Audacity-X.X.X", CMake cannot parse the
returned string from git and fails.

As fix, pass the argument "--match" with a glob pattern to the
"git describe" call so that it looks for the latest tag matching the
expected pattern. This should also fix the sr.ht CI builds.

Al suggested by review, match the generic pattern "*.*.*" which matches
a generic tag like "1.0.0" without any prefix as well as "v1.0.0" or
"1.0.0-rc1". A better filter that e.g. validates that only number are
between the dots seems to be impossible with the glob syntax of
"git describe".

Reference-to: https://github.com/tenacityteam/tenacity/issues/640.

Signed-off-by: Max Maisel <max.maisel@posteo.de>